Writing tests, eliminating answer_checker class, minor changes to game class
