+++ top_ten.yml
@@ -4,3 +4,33 @@ name and answer:
 - PPPGRRRB
 time: 7
 number of guesses: 1
+---
+name and answer:
+- Judson
+- PPGRGR
+time: 258
+number of guesses: 4
+---
+name and answer:
+- Judson
+- BYRY
+time: 8
+number of guesses: 2
+---
+name and answer:
+- Judson
+- BBRR
+time: 13
+number of guesses: 2
+---
+name and answer:
+- KK
+- YBBY
+time: 5
+number of guesses: 2
+---
+name and answer:
+- KK
+- GRYYRR
+time: 9
+number of guesses: 1
